refactor(MBeanTree): refactored code so both JMX and camel components have the same search behaviour
